Add _joinData prop for HABTM.
